requirements/common.txt
amqp==1.4.9
anyjson==0.3.3
appdeps==1.0.0
APScheduler==3.0.2
Django==1.8.8
django-allauth==0.24.1
django-braces==1.8.1
django-crispy-forms==1.5.2
django-jsonfield==0.9.15
funcsigs==0.4
futures==3.0.4
iso8601==0.1.11
kombu==3.0.29
librabbitmq==1.6.1
meld3==1.0.2
mock==1.3.0
oauthlib==1.0.3
pbr==1.8.1
psycopg2==2.6.1
python-openid==2.2.5
pytz==2015.7
requests==2.9.1
requests-oauthlib==0.6.0
six==1.10.0
SQLAlchemy==1.0.11
supervisor==3.2.0
tzlocal==1.2
urllib3==1.14
warc==0.2.1
djangorestframework==3.3.2
django-filter==0.12.0